Extensions Refactor
