+++ src/typechat.program/TypeChat.Program.csproj
@@ -21,6 +21,7 @@
 
   <ItemGroup>
     <Compile Include="..\typechat\ArgumentVerify.cs" Link="ArgumentVerify.cs" />
+    <Compile Include="..\typechat\ListExtensions.cs" Link="ListExtensions.cs" />
   </ItemGroup>
 
   <ItemGroup>
