+++ src/typechat.schema/TypeChat.Schema.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;net60;net70</TargetFrameworks>
@@ -11,6 +11,8 @@
 
   <ItemGroup>
     <Compile Include="..\typechat\ArgumentVerify.cs" Link="ArgumentVerify.cs" />
+    <Compile Include="..\typechat\StringExtensions.cs" Link="StringExtensions.cs" />
+    <Compile Include="..\typechat\ListExtensions.cs" Link="ListExtensions.cs" />
   </ItemGroup>
   
   <ItemGroup>
